Make purple events fire again
6da78ec2d930c8612cf4fe302a469c0ddf358bb5 made the cooldown interval
between purple events configurable in mod settings, but it ended up
making purple events almost never fire, because it sets
`tickLastPurpleEvent` in `CanFireNowSub()`, which is called not just
when an incident is about to fire, but also when the storyteller is
generating a list of potential incidents to fire. Fix it by setting
`tickLastPurpleEvent` in `TryExecuteWorker` instead, which is when the
event actually fires.